requirements.txt
Django==2.1
django-debug-toolbar==1.11
django-split-settings==0.3.0
django-widget-tweaks==1.4.3
isort==4.3.20
Pillow==6.0.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
